% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_conda.R
\name{ActivateConda}
\alias{ActivateConda}
\title{Activate Conda Environment}
\usage{
ActivateConda(conda.env = "rgee")
}
\arguments{
\item{conda.env}{[Optional] Conda environment. If \code{NULL} is selected,
the user can choose the environment from a list. Default value is
\code{"rgee"}.}
}
\description{
This function provides users with the flexibility to either
specify a Conda environment or choose from a list of available environments
if none is explicitly defined.
}
\examples{
# Example 1: Activate Conda environment without providing an environment name
\dontrun{
ActivateConda()
}

# Example 2: Activate Conda environment with providing an environment name
\dontrun{
ActivateConda(conda.env = "rgee")
}

}
