% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_drive_folders_by_name}
\alias{clean_drive_folders_by_name}
\title{Clean Contents or Entire Google Drive Folders by Name}
\usage{
clean_drive_folders_by_name(
  folder_name,
  delete_folders = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{folder_name}{[mandatory] (character) Name of the folder(s) to search
for in Google Drive.}

\item{delete_folders}{[optional] (logical) If \code{TRUE}, the matched
folders themselves will be deleted after their contents are removed.
Default is \code{FALSE}.}

\item{verbose}{[optional] (logical) If \code{TRUE}, messages will be printed
about cleanup actions. Default is \code{TRUE}.}
}
\description{
Searches for all Google Drive folders with the specified name and optionally
removes their contents and/or the folders themselves. Useful for cleaning up
scratch or export folders used by Earth Engine batch processes.
}
\keyword{internal}
