% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{compare_lists}
\alias{compare_lists}
\title{Compare Lists and Highlight Differences}
\usage{
compare_lists(list_1, list_2)
}
\arguments{
\item{list_1}{[mandatory] (list) First list to compare.}

\item{list_2}{[mandatory] (list) Second list to compare.}
}
\value{
A list showing added and removed values marked with '+' and '-'.
}
\description{
Compares two lists and mark new values with '+' and removed values with '-'.
}
\keyword{internal}
