% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{compare_vectors}
\alias{compare_vectors}
\title{Compare Vectors and Highlight Differences}
\usage{
compare_vectors(vector_1, vector_2)
}
\arguments{
\item{vector_1}{[mandatory] (character or integer) First vector to compare.}

\item{vector_2}{[mandatory] (character or integer) Second vector to compare.}
}
\value{
A vector showing added and removed values marked with '+' and '-'.
}
\description{
Compares two vectors and indicates added ('+') and removed ('-') values.
}
\keyword{internal}
