% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{drive_auth_ready}
\alias{drive_auth_ready}
\title{Check and Initialize Google Drive Authentication}
\usage{
drive_auth_ready(
  email = NULL,
  cache_path = NULL,
  interactive_fallback = interactive()
)
}
\arguments{
\item{email}{[optional] (character) Email to target.}

\item{cache_path}{[optional] (character) Directory to search for tokens.}

\item{interactive_fallback}{[optional] (logical) Allow interactive auth.}
}
\description{
Checks for an existing token in the specified cache path.
}
\keyword{internal}
