% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R,
%   R/run_geelite._halfgoogleimportR.R
\name{ee_as_tidyee}
\alias{ee_as_tidyee}
\title{Minimal Replacement for tidyrgee::as_tidyee}
\usage{
ee_as_tidyee(ic)

ee_as_tidyee(ic)
}
\arguments{
\item{ic}{[mandatory] (ee$ImageCollection) The Earth Engine image collection
to convert.}
}
\value{
A list with:
  \code{ee_ob} (the original ImageCollection) and
  \code{vrt} (list containing a \code{date} vector).

A list with:
  \code{ee_ob} (the original ImageCollection) and
  \code{vrt} (list containing a \code{date} vector).
}
\description{
Creates a lightweight representation of an Earth Engine
\code{ee$ImageCollection}, containing the original collection and its
associated dates extracted from \code{system:time_start}.

Creates a lightweight representation of an Earth Engine
\code{ee$ImageCollection}, containing the original collection and its
associated dates extracted from \code{system:time_start}.
}
\keyword{internal}
