% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{expand_to_daily}
\alias{expand_to_daily}
\title{Expand Data to Daily Frequency}
\usage{
expand_to_daily(df_long, prep_fun)
}
\arguments{
\item{df_long}{[mandatory] (data.frame) A long-format data frame with at
least the columns \code{id} and \code{date}.}

\item{prep_fun}{[mandatory] (function) Function used for pre-processing.}
}
\value{
A data frame with daily dates and preprocessed \code{value} column.
}
\description{
Expands the input data frame to a daily frequency, filling in any missing
dates within the observed range.
}
\keyword{internal}
