% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{extract_batch_stat}
\alias{extract_batch_stat}
\title{Extract Grid Statistics for a Batch of Bins}
\usage{
extract_batch_stat(imgs, grid, dates, band, stat, stat_fun, scale)
}
\arguments{
\item{imgs}{[mandatory] (ImageCollection) GEE object containing images.}

\item{grid}{[mandatory] (sf) Simple features object containing the
geometries of the regions of interest.}

\item{dates}{[mandatory] (date) Dates when the targeted GEE dataset is
available, defining the time span for data extraction.}

\item{band}{[mandatory] (character) Name of the band to extract statistics
from.}

\item{stat}{[mandatory] (character) Name of the statistical measure applied
for grid statistics calculation.}

\item{stat_fun}{[mandatory] (function) Function used to calculate the grid
statistic.}

\item{scale}{[mandatory] (numeric) Scale of images before processing.}
}
\value{
A data frame containing grid statistics for the specified batch of
bins.
}
\description{
Extracts grid statistics from Google Earth Engine (GEE) images for a batch
of bins.
}
\keyword{internal}
