% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{extract_grid_stats}
\alias{extract_grid_stats}
\title{Extract Grid Statistics}
\usage{
extract_grid_stats(
  case,
  images,
  grid,
  batch_1,
  batch_2,
  band,
  stat,
  stat_fun,
  scale,
  region_new,
  skip_update,
  grid_stats
)
}
\arguments{
\item{case}{[mandatory] (logical) Type of data collection request (\code{1}:
Build, \code{2}: Update).}

\item{images}{[mandatory] (list) List containing GEE images and related
information.}

\item{grid}{[mandatory] (sf) Simple features object containing the
geometries of the regions of interest.}

\item{batch_1}{[mandatory] (integer) Batch from primary batches.}

\item{batch_2}{[mandatory] (integer) Batch from additional batches.}

\item{band}{[mandatory] (character) Name of the band to extract statistics
from.}

\item{stat}{[mandatory] (character) Name of the statistical measure applied
for grid statistics calculation.}

\item{stat_fun}{[mandatory] (function) Function used to calculate the grid
statistic.}

\item{scale}{[mandatory] (numeric) Scale of images before processing.}

\item{region_new}{[mandatory] (logical) Vector indicating whether new
regions are to be processed.}

\item{skip_update}{[mandatory] (logical) Logical value indicating whether
the 'band' is up-to-date.}

\item{grid_stats}{[mandatory] (list) A list containing previously calculated
grid statistics.}
}
\value{
A data frame containing grid statistics for both building and
updating processes.
}
\description{
Extracts grid statistics from Google Earth Engine (GEE) images based on
specified parameters.
}
\keyword{internal}
