% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_regions.R
\name{fetch_country_regions}
\alias{fetch_country_regions}
\title{Fetch Country-Level Regions}
\usage{
fetch_country_regions()
}
\value{
A data frame of country-level regions.
}
\description{
Retrieves country-level ISO 3166-2 regions.
}
\keyword{internal}
