% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_regions.R
\name{fetch_state_regions}
\alias{fetch_state_regions}
\title{Fetch State-Level Regions}
\usage{
fetch_state_regions()
}
\value{
A data frame containing state-level region names and ISO 3166-2
  codes.
}
\description{
Retrieves ISO 3166-2 codes for state-level regions.
}
\keyword{internal}
