% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{fetch_vars}
\alias{fetch_vars}
\title{Fetch Variable Information from an SQLite Database}
\usage{
fetch_vars(path, print_output = TRUE)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}

\item{print_output}{[optional] (logical) If \code{TRUE}, prints the output
in a markdown format; if \code{FALSE}, returns a data frame object (default:
\code{TRUE}).}
}
\value{
If \code{print_output = TRUE}, the function prints the variable
information in a markdown format. If \code{print_output = FALSE}, it returns
a data frame object.
}
\description{
Displays information on the available variables in the SQLite database
(\code{data/geelite.db}).
}
\examples{
# Example: Printing the available variables
\dontrun{
  fetch_vars(path = "path/to/db")
}
}
