% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{filter_and_sort_tables}
\alias{filter_and_sort_tables}
\title{Filter and Sort Tables}
\usage{
filter_and_sort_tables(tables)
}
\arguments{
\item{tables}{[mandatory] (character or integer) A vector of table names.}
}
\value{
A data frame with filtered and sorted table names.
}
\description{
Filters out system tables and sorts the table names.
}
\keyword{internal}
