% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{gee_message}
\alias{gee_message}
\title{Print Google Earth Engine and Python Environment Information}
\usage{
gee_message(user)
}
\arguments{
\item{user}{[mandatory] (character) Used to create a directory within the
path \code{~/.config/earthengine/}. This directory stores all the
credentials associated with a specific Google account.}
}
\description{
This function prints out information related to the Google Earth Engine
(GEE) and Python environment setup.
}
\keyword{internal}
