% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R,
%   R/run_geelite._halfgoogleimportR.R
\name{gen_messages}
\alias{gen_messages}
\title{Define Output Messages}
\usage{
gen_messages(database_new)

gen_messages(database_new)
}
\arguments{
\item{database_new}{[mandatory] (logical) A logical value indicating whether
the database is new.}
}
\value{
A list of output messages.

A list of output messages.
}
\description{
Defines output messages based on whether the database is new or updated.

Defines output messages based on whether the database is new or updated.
}
\keyword{internal}
