% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_config.R
\name{GenerateConfig}
\alias{GenerateConfig}
\title{Configuration File Generation}
\usage{
GenerateConfig(target.path = NULL)
}
\arguments{
\item{target.path}{[Optional] Path to the directory. If not provided, the
user is prompted to enter a path. Default value is \code{NULL}.}
}
\description{
This function sets up the configuration file for data generation
and update.
}
\examples{
# Example 1: Run configuration without providing a target path
\dontrun{
  GenerateConfig()
}

# Example 2: Run configuration with a target path
\dontrun{
  GenerateConfig(target.path = "/path/to/custom/directory")
}

}
