% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_path.R
\name{GeneratePath}
\alias{GeneratePath}
\title{Custom Target Path Generation}
\usage{
GeneratePath()
}
\value{
The selected \code{target.path}.
}
\description{
This function prompts the user to enter a valid target path. If
the user cancels the input or provides an invalid path, the function will ask
again until a valid path is provided.
}
\examples{
# Example: Run custom path generation
\dontrun{
result <- GeneratePath()
print(result)
}

}
