% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_task.R
\name{GenerateTask}
\alias{GenerateTask}
\title{Task Generation}
\usage{
GenerateTask(target.path = NULL, operation.code = NULL)
}
\arguments{
\item{target.path}{[Optional] Path to the directory. If not provided, the
user is prompted to enter a path. Default value is \code{NULL}.}

\item{operation.code}{[Optional] For updating, use \code{1}, and for
(re)building, use \code{2}. If not provided, the user is prompted to enter a
path. The default value is \code{NULL}.}
}
\value{
A list containing the \code{task} and \code{post.state} objects.
}
\description{
This function generates the task that controls the updating or
(re)building operation. Additionally, it creates the \code{post.state}
object. Upon successful (re)building/update, the \code{state.json} file is
updated with information from this \code{post.state} object.
}
\examples{
# Example 1: Run task generation without an operation code and a target path
\dontrun{
result  <- GenerateTask()
print(result)
}

# Example 2: Run task generation with an operation code and a target path
\dontrun{
result <- GenerateTask(target.path = "/path/to/custom/directory",
operation.code = 2)
}

}
