% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{get_batch}
\alias{get_batch}
\title{Divide Grid Into Batches}
\usage{
get_batch(grid, batch_size = NULL, batch_num = NULL)
}
\arguments{
\item{grid}{[mandatory] (sf) Simple features object containing the
geometries of the regions of interest.}

\item{batch_size}{[optional] (integer) Maximum size of each batch
(default: \code{NULL}). If \code{NULL}, 'batch_num' must be specified.}

\item{batch_num}{[optional] (integer) Number of batches (default:
\code{NULL}). If \code{NULL}, 'batch_size' must be specified.}
}
\value{
A list containing the batches.
}
\description{
Divides the 'grid' object into batches based on either the specified batch
size or the number of batches.
}
\keyword{internal}
