% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R,
%   R/run_geelite._halfgoogleimportR.R
\name{get_cases}
\alias{get_cases}
\title{Determine the Cases of Data Collection Requests}
\usage{
get_cases(database_new, dataset_new, band_new, stats_new, regions_new)

get_cases(database_new, dataset_new, band_new, stats_new, regions_new)
}
\arguments{
\item{database_new}{[mandatory] (logical) A logical value indicating whether
the database is new.}

\item{dataset_new}{[mandatory] (logical) A logical value indicating whether
the dataset is new.}

\item{band_new}{[mandatory] (logical) A logical value indicating whether the
band is new.}

\item{stats_new}{[mandatory] (logical) A logical vector indicating which
statistics are new.}

\item{regions_new}{[mandatory] (logical) A logical vector indicating which
regions are new.}
}
\value{
An integer indicating the processing cases as follows:
\describe{
 \item{1}{All build}
 \item{2}{All update}
 \item{3}{Mixed}
}

An integer indicating the processing cases as follows:
\describe{
 \item{1}{All build}
 \item{2}{All update}
 \item{3}{Mixed}
}
}
\description{
Determines the cases of data collection requests based on the markers of
'datasets', 'bands', and 'stats'.

Determines the cases of data collection requests based on the markers of
'datasets', 'bands', and 'stats'.
}
\keyword{internal}
