% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{get_config}
\alias{get_config}
\title{Printing the Configuration File}
\usage{
get_config(path)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}
}
\value{
Configuration object.
}
\description{
This function prints the configuration file.
}
\examples{
# Example: Printing the configuration file
\dontrun{
  get_config(path = "path")
}

}
