% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fips.R
\name{get_fips}
\alias{get_fips}
\title{Collecting FIPS Region Codes}
\usage{
get_fips(type = "country")
}
\arguments{
\item{type}{[optional] (character) Type of the regions to be printed
(options: \code{"state"}, \code{"country"}, \code{"all"}, default:
\code{"country"}).}
}
\value{
A data frame object that includes region names, FIPS codes, and
types.
}
\description{
This function collects FIPS region codes.
}
\examples{
# Example: Storing FIPS region codes
\dontrun{
 fips <- get_fips()
}

}
