% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R,
%   R/run_geelite._halfgoogleimportR.R
\name{get_grid}
\alias{get_grid}
\title{Obtain H3 Hexagonal Grid}
\usage{
get_grid(task)

get_grid(task)
}
\arguments{
\item{task}{[mandatory] (list) Session task that specifies the parameters
for data collection.}
}
\value{
A simple features (sf) object containing grid data.

A simple features (sf) object containing grid data.
}
\description{
Retrieves or creates the grid for the task based on the specified regions
and resolution.

Retrieves or creates the grid for the task based on the specified regions
and resolution.
}
\keyword{internal}
