% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{get_reducers}
\alias{get_reducers}
\title{Get Reducers}
\usage{
get_reducers()
}
\value{
A list of available reducers.
}
\description{
Initializes a list of reducers for grid statistics calculation.
}
\keyword{internal}
