% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regids.R
\name{get_regids}
\alias{get_regids}
\title{Collecting ISO 3166-2 Region Codes}
\usage{
get_regids(type = "country")
}
\arguments{
\item{type}{[optional] (character) Type of the regions to be printed
(options: \code{"state"}, \code{"country"}, \code{"all"}, default:
\code{"country"}).}
}
\value{
A data frame object that includes region names, ISO 3166-2 codes, and
types.
}
\description{
This function collects ISO 3166-2 region codes.
}
\examples{
# Example: Storing ISO 3166-2 region codes
\dontrun{
 regids <- get_regids()
}

}
