% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{get_shapes}
\alias{get_shapes}
\title{Get Shapes for Specified Regions}
\usage{
get_shapes(regions)
}
\arguments{
\item{regions}{[mandatory] (character) A vector containing ISO 3166-2 region
codes. Country codes are two characters long, while state codes contain
additional characters.}
}
\value{
A simple features (sf) object containing the shapes of the specified
  regions.
}
\description{
Retrieves the shapes of specified regions, which can be at the country or
state level.
}
\keyword{internal}
