% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state.R
\name{get_state}
\alias{get_state}
\title{Printing the State File}
\usage{
get_state(path)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}
}
\value{
State object.
}
\description{
This function prints the state file.
}
\examples{
# Example: Printing the state file
\dontrun{
  get_state(path = "path")
}

}
