% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{get_tables}
\alias{get_tables}
\title{Printing the Names of the Available SQLite Tables}
\usage{
get_tables(path)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}
}
\value{
Table names.
}
\description{
This function prints the names of the available SQLite tables.
}
\examples{
# Example: Printing the available SQLite tables
\dontrun{
  get_tables(path = "path")
}

}
