% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R,
%   R/run_geelite._halfgoogleimportR.R
\name{get_task}
\alias{get_task}
\title{Generate Session Task}
\usage{
get_task()

get_task()
}
\value{
A list representing the session task.

A list representing the session task.
}
\description{
Generates a session task based on the configuration and state files.

Generates a session task based on the configuration and state files.
}
\keyword{internal}
