% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gee_data.R
\name{GetGEEData}
\alias{GetGEEData}
\title{Collecting Google Earth Engine Data}
\usage{
GetGEEData(task, resolution = 4, scale = 10000)
}
\arguments{
\item{task}{[Mandatory] This object, generated by the \code{GenerateTask}
function, manages the update or (re)building operation.}

\item{resolution}{[Optional] The resolution for hexagons. Default value is
\code{4}.}

\item{scale}{[Optional] The scale for Earth Engine data extraction. Default
value is \code{10000}.}
}
\value{
An \code{sf} object containing the collected Google Earth Engine
data.
}
\description{
This function collects data from Google Earth Engine based on
the specified task.
}
\examples{
# Example: Retrieve Google Earth Engine data using a task file
\dontrun{
result <- GetGEEData(task = task)
print(result)
}

}
