% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{init_postp}
\alias{init_postp}
\title{Initialize Post-Processing Folder and Files}
\usage{
init_postp(path, verbose = TRUE)
}
\arguments{
\item{path}{[mandatory] \code{character} The path to the root directory
where the \code{postp} folder should be created.}

\item{verbose}{[optional] (logical) Display messages (default: \code{TRUE}).}
}
\value{
No return value, called for side effects.
}
\description{
Creates a \code{postp} folder at \code{path} and adds two files:
\code{structure.json} and \code{functions.R}.
}
\details{
The \code{structure.json} file is initialized with \code{"default": null}.
The \code{functions.R} file contains simple example functions that match the
default "do nothing" behavior plus a z-score example.
}
\examples{
# Example: Initialize post-processing files in the database directory
\dontrun{
  init_postp("path/to/db")
}

}
