% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{linear_interp}
\alias{linear_interp}
\title{Simple Linear Interpolation}
\usage{
linear_interp(x)
}
\arguments{
\item{x}{[mandatory] (numeric) A numeric vector possibly containing
\code{NA} values.}
}
\value{
A numeric vector of the same length as \code{x}, with \code{NA}
  values replaced by linear interpolation.
}
\description{
Replaces \code{NA} values with linear interpolation.
}
\keyword{internal}
