% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_config.R
\name{mod_config}
\alias{mod_config}
\title{Modifying Configuration File}
\usage{
mod_config(path, target, values)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}

\item{target}{[mandatory] (list) Specifies the path to the values to be
replaced.}

\item{values}{[mandatory] (list) New values to replace the original ones.}
}
\description{
This function modifies the configuration file.
}
\examples{
# Example: Modify configuration file
\dontrun{
mod_config(path = "/path/to/custom/directory",
           list("limit", c("source", "MODIS/006/MOD13A2", "NDVI")),
}

}
