% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{output_message}
\alias{output_message}
\title{Output Message}
\usage{
output_message(message, verbose)
}
\arguments{
\item{message}{[mandatory] (list) The message to be displayed.}

\item{verbose}{[mandatory] (logical) Flag indicating whether to display the
message.}
}
\description{
Outputs message if verbose mode is \code{TRUE}.
}
\keyword{internal}
