% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R,
%   R/run_geelite._halfgoogleimportR.R
\name{print_version}
\alias{print_version}
\title{Display geeLite Package Version}
\usage{
print_version(verbose)

print_version(verbose)
}
\arguments{
\item{verbose}{[mandatory] (logical) If \code{TRUE}, the version of the
\code{geeLite} package is printed.}
}
\description{
Displays the version of the \code{geeLite} package with formatted headers.

Displays the version of the \code{geeLite} package with formatted headers.
}
\keyword{internal}
