% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cli.R
\name{process_single_file}
\alias{process_single_file}
\title{Process a Single Source File}
\usage{
process_single_file(src_file_path, path)
}
\arguments{
\item{src_file_path}{[mandatory] (character) Path of the source file.}

\item{path}{[mandatory] (character) The path to the root directory of the
generated database.}
}
\description{
Processes an individual source file by specifying the 'path' within the
script, and writes the output to the CLI directory of the database.
}
\keyword{internal}
