% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cli.R
\name{process_source_files}
\alias{process_source_files}
\title{Process Source Files}
\usage{
process_source_files(src_files_path, path)
}
\arguments{
\item{src_files_path}{[mandatory] (character) A vector of source file paths.}

\item{path}{[mandatory] (character) The path to the root directory of the
generated database.}
}
\description{
Processes multiple source files by iterating through them.
}
\keyword{internal}
