% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{process_vector}
\alias{process_vector}
\title{Process Marked Vector}
\usage{
process_vector(vector)
}
\arguments{
\item{vector}{[mandatory] (character) A character vector containing elements
marked with '+' and '-' prefixes.}
}
\value{
A list with the following components:
\describe{
 \item{$add}{Items marked with '+'}
 \item{$drop}{Items marked with '-'}
 \item{$use}{Items that are unmarked or marked with '+'}
 \item{$use_add}{TRUE for items marked with '+' within the $use category}
}
}
\description{
Generates a list categorizing items based on their marks: items to be added
('+'), items to be dropped ('-'), items to be used (unmarked or marked with
'+'), and indices of '+' items within the used category.
}
\keyword{internal}
