% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{read_grid}
\alias{read_grid}
\title{Read Grid from Database}
\usage{
read_grid()
}
\value{
A simple features (sf) object containing the grid data.
}
\description{
Reads the H3 grid from the specified SQLite database
(\code{data/geelite.db}).
}
\keyword{internal}
