% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{read_tables}
\alias{read_tables}
\title{Read Tables from Database}
\usage{
read_tables(path, tables)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}

\item{tables}{[mandatory] (character) A vector of table names to be read.}
}
\value{
A list of tables read from the database.
}
\description{
Reads the specified tables from the SQLite database.
}
\keyword{internal}
