% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{read_variables}
\alias{read_variables}
\title{Read Variables from Database}
\usage{
read_variables(path, variables, freq, prep_fun, aggr_funs, postp_funs)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}

\item{variables}{[mandatory] (character) A vector of variable names to read.}

\item{freq}{[mandatory] (character) Specifies the frequency to aggregate the
data.}

\item{prep_fun}{[mandatory] (function) Function used for pre-processing.}

\item{aggr_funs}{[mandatory] (function or list) Aggregation function(s).}

\item{postp_funs}{[mandatory] (function or list) Post-processing function(s).}
}
\value{
A list of variables read from the database.
}
\description{
Reads the specified variables from the SQLite database.
}
\keyword{internal}
