% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{run_geelite}
\alias{run_geelite}
\title{Collecting and Storing GEE Data}
\usage{
run_geelite(path, conda = "rgee", user = NULL, rebuild = FALSE, verbose = TRUE)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}

\item{conda}{[optional] (character) Name of the virtual Conda environment
installed and used by the rgee package (default: \code{"rgee"}).}

\item{user}{[optional] (character) Used to create a folder within the path
\code{~/.config/earthengine/}. This folder stores all the credentials
associated with a specific Google identity (default: \code{NULL}).}

\item{rebuild}{[optional] (logical) If set to \code{TRUE}, the database and
its supplementary files will be overwritten based on the configuration file
(default: \code{FALSE}).}

\item{verbose}{[optional] (logical) Display computation status and messages
(default: \code{TRUE}).}
}
\description{
Collects and stores GEE data, updating states and log files
(\code{path/data/geelite.db}).
}
\examples{
# Example: Data collection
\dontrun{
  run_geelite(path = "path")
}

}
