% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gfi_data.R
\name{RunGFIData}
\alias{RunGFIData}
\title{Collecting and Updating Global Food Insecurity Data}
\usage{
RunGFIData(
  target.path = NULL,
  operation.code = NULL,
  conda.env = "rgee",
  quiet = FALSE
)
}
\arguments{
\item{target.path}{[Optional] Path to the directory. If not provided, the
user is prompted to enter a path. Default value is \code{NULL}.}

\item{operation.code}{[Optional] For updating, use \code{1}, and for
(re)building, use \code{2}. If not provided, the user is prompted to enter a
path. The default value is \code{NULL}.}

\item{conda.env}{[Optional] Conda environment. If \code{NULL} is selected,
the user can choose the environment from a list. Default value is
\code{"rgee"}.}

\item{quiet}{[Optional] If \code{TRUE}, suppresses prompts or user
interactions. Default value is \code{FALSE}.}
}
\description{
This function performs the entire update or (re)build operation.
}
\examples{
# Example 1: Run update or (re)build operation without providing a custom
# path
\dontrun{
  RunGFIData()
}

# Example 2: Run update or (re)build operation with a custom path
\dontrun{
  RunGFIData(target.path = "/path/to/custom/directory", conda.env = "rgee")
}

}
