% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cli.R
\name{set_cli}
\alias{set_cli}
\title{Initialize CLI Files}
\usage{
set_cli(path, verbose = TRUE)
}
\arguments{
\item{path}{[mandatory] (character) The path to the root directory of the
generated database.}

\item{verbose}{[optional] (logical) Whether to display messages (default:
\code{TRUE}).}
}
\description{
Creates R scripts to enable the main functions to be called through the
Command Line Interface (CLI). These scripts are stored in the \code{cli/}
directory of the generated database.
}
\examples{
# Example: Setting up CLI files
\dontrun{
  set_cli(path = "path/to/db")
}

}
