% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{set_dirs}
\alias{set_dirs}
\title{Generate Necessary Directories}
\usage{
set_dirs(rebuild)
}
\arguments{
\item{rebuild}{[optional] (logical) If \code{TRUE}, existing directories
will be removed and recreated.}
}
\description{
Generates \code{"data"}, \code{"log"}, \code{"cli"}, and \code{"state"}
subdirectories at the specified path.
}
\keyword{internal}
