% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{set_progress_bar}
\alias{set_progress_bar}
\title{Set Progress Bar}
\usage{
set_progress_bar(verbose)
}
\arguments{
\item{verbose}{[mandatory] (logical) If \code{TRUE}, a progress bar is
initialized.}
}
\value{
A progress bar (environment) if 'verbose' is \code{TRUE}, or
\code{NULL} if \code{FALSE}.
}
\description{
Initializes a progress bar if 'verbose' is \code{TRUE}.
}
\keyword{internal}
