% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{update_grid_stats}
\alias{update_grid_stats}
\title{Update Grid Statistics}
\usage{
update_grid_stats(grid_stat, batch_stat)
}
\arguments{
\item{grid_stat}{[optional] (data.frame) Existing data frame of grid
statistics to append the newly calculated statistics to.}

\item{batch_stat}{[mandatory] (data.frame) New data frame of grid statistics
to append to the existing statistics.}
}
\value{
(data.frame) A combined data frame with missing columns filled as NA.

A data frame containing the updated grid statistics.
}
\description{
Updates existing grid statistics with newly calculated statistics.
}
\keyword{internal}
