% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_params}
\alias{validate_params}
\title{Validate Parameters}
\usage{
validate_params(params)
}
\arguments{
\item{params}{[mandatory] (list) Parameters to be validated.}
}
\value{
Invisible \code{NULL} if all validations pass.
}
\description{
Validates multiple parameters.
}
\details{
Validations performed:
- 'admin_lvl': Validates if it is \code{NULL}, \code{0}, or \code{1}.
- 'conda': Checks if it is an available Conda environment.
- 'file_path': Constructs a file path and checks if the file exists.
- 'keys': Ensures it is a non-empty list with valid entries.
- 'limit': Checks if it is a positive numeric value.
- 'new_values': Verifies it is a list with the same length as 'keys'.
- 'user': Checks if it is \code{NULL} or a character value.
- 'path': Checks if the directory exists.
- 'print_output': Checks if it is a logical value.
- 'rebuild': Checks if it is a logical value.
- 'regions': Checks if the first two characters are letters.
- 'start': Checks if it is a valid date.
- 'verbose': Checks if it is a logical value.
}
\keyword{internal}
