% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_source_param}
\alias{validate_source_param}
\title{Validate Source Parameter}
\usage{
validate_source_param(source)
}
\arguments{
\item{source}{[mandatory] (list) A list containing datasets, each with its
associated bands and statistics. The structure should follow a nested
format where each dataset is a named list, each band within a dataset is
also a named list, and each statistic within a band is a non-empty
character string.}
}
\value{
Returns \code{TRUE} if the 'source' parameter is valid. Throws an
  error if the parameter is invalid.
}
\description{
Checks the validity of the 'source' parameter.
}
\keyword{internal}
