% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R,
%   R/run_geelite._halfgoogleimportR.R
\name{write_grid_stats}
\alias{write_grid_stats}
\title{Write Grid Statistics to Database}
\usage{
write_grid_stats(database_new, dataset_new, dataset, db_table, grid_stats)

write_grid_stats(database_new, dataset_new, dataset, db_table, grid_stats)
}
\arguments{
\item{database_new}{[mandatory] (logical) A logical value indicating whether
the database is new.}

\item{dataset_new}{[mandatory] (logical) A logical value indicating whether
the dataset is new.}

\item{dataset}{[mandatory] (character) Name of the dataset to initialize or
update in the SQLite database.}

\item{db_table}{[mandatory] (data.frame) The table to be updated or
retrieved from the SQLite database. Set to \code{NULL} during the
(re)building procedure.}

\item{grid_stats}{[mandatory] (list) List containing grid statistics
separately for (re)building and updating procedures.}
}
\description{
Writes grid statistics to the SQLite database.

Writes grid statistics to the SQLite database.
}
\keyword{internal}
