% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R,
%   R/run_geelite._halfgoogleimportR.R
\name{write_log_file}
\alias{write_log_file}
\title{Write Log File}
\usage{
write_log_file(database_new)

write_log_file(database_new)
}
\arguments{
\item{database_new}{[mandatory] (logical) A logical value indicating whether
the database is new.}
}
\description{
Writes the log file to the specified directory within the generated
database (\code{log/log.txt}).

Writes the log file to the specified directory within the generated
database (\code{log/log.txt}).
}
\keyword{internal}
