% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_sqlite.R
\name{WriteToSQLite}
\alias{WriteToSQLite}
\title{Writing Data to SQLite Database}
\usage{
WriteToSQLite(data, table.name, operation.code, target.path = NULL)
}
\arguments{
\item{data}{[Mandatory] The data to be written to the SQLite database.}

\item{table.name}{[Mandatory] The name of the table to write the data into.}

\item{operation.code}{[Mandatory] For updating, use \code{1}, and for
(re)building, use \code{2}. If not provided, the user is prompted to enter a
path.}

\item{target.path}{[Optional] Path to the directory. If not provided, the
user is prompted to enter a path. Default value is \code{NULL}.}
}
\description{
This function writes data to an SQLite database and creates it
if it does not exist. It connects to the specified SQLite database file,
creates a new table if it does not exist, and appends new values to the
existing table if it does.
}
\examples{
# Example: Update 'my_table' in 'my.data' SQLite database
\dontrun{
WriteToSQLite(data = my.data, table.name = "my_table", operation.code = 1)
}

}
