% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{aggr_by_freq}
\alias{aggr_by_freq}
\title{Aggregate Data by Frequency}
\usage{
aggr_by_freq(table, freq, funs)
}
\arguments{
\item{table}{[mandatory] (data.frame) A wide-format data frame object of the
generated SQLite table.}

\item{freq}{[mandatory] (character) Specifies the frequency to aggregate the
data (options: \code{"month"}, \code{"year"}).}

\item{funs}{[optional] (character) A character vector of statistical
functions aggregation is based on (options: \code{NULL}, \code{"mean"},
\code{"median"}, \code{"min"}, \code{"max"}, \code{"sd"}).}
}
\value{
A data frame in wide format where columns represent aggregated values
for each frequency period and statistic applied.
}
\description{
This function aggregates data from a wide-format data frame according to a
specified frequency and selected statistical functions.
}
\keyword{internal}
