% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{expand_to_daily}
\alias{expand_to_daily}
\title{Expand Data to Daily Frequency}
\usage{
expand_to_daily(df_long, prep_fun)
}
\arguments{
\item{df_long}{[mandatory] (data.frame) A long-format data frame with at
least the columns \code{id}, \code{band}, \code{zonal_stat}, and \code{date}.}

\item{prep_fun}{[mandatory] (function) A single function used for
pre-processing the time series data before aggregation. This function
converts the data to a daily frequency and applies any necessary data
transformation or imputation.}
}
\value{
A data frame with daily dates and preprocessed \code{value} column.
}
\description{
This function expands the input data frame (\code{df_long}) to a daily
frequency, filling in any missing dates within the observed range. A
preprocessing function (\code{prep_fun}) is applied to handle missing or
interpolated values.
}
\keyword{internal}
