% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_regions.R
\name{fetch_country_regions}
\alias{fetch_country_regions}
\title{Fetch Country-Level Regions}
\usage{
fetch_country_regions()
}
\value{
A data frame containing country-level region names and ISO 3166-2
  codes.
}
\description{
Retrieves ISO 3166-2 codes for country-level regions.
}
\keyword{internal}
