% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_regions.R
\name{fetch_state_regions}
\alias{fetch_state_regions}
\title{Fetch State-Level Regions}
\usage{
fetch_state_regions()
}
\value{
A data frame of state-level regions.
}
\description{
Retrieves state-level ISO 3166-2 regions.
}
\keyword{internal}
