% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_regions.R
\name{fetch_state_regions}
\alias{fetch_state_regions}
\title{Fetch ISO 3166-2 Subdivision Codes}
\usage{
fetch_state_regions()
}
\value{
A data frame with subdivision names, ISO 3166-2 codes, and admin
  level 1.
}
\description{
Retrieves first-level administrative subdivisions (e.g., states, provinces)
using ISO 3166-2 codes.
}
\keyword{internal}
