% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{fetch_tables}
\alias{fetch_tables}
\title{Fetch Names of SQLite Tables}
\usage{
fetch_tables(path)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}
}
\value{
A data frame containing the table names.
}
\description{
Reads and prints the names of available SQLite tables in the specified
directory of the generated database (\code{data/geelite.db}).
}
\examples{
# Example: Printing the names of available SQLite tables
\dontrun{
  fetch_tables(path = "path/to/db")
}
}
