% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_install.R
\name{gee_install}
\alias{gee_install}
\title{Create a Conda Environment with 'rgee' Dependencies}
\usage{
gee_install(conda = "rgee")
}
\arguments{
\item{conda}{[optional] (character) The name of the virtual Conda
environment used by the \code{rgee} package (default: \code{"rgee"}).}
}
\description{
Sets up an isolated Conda environment with all the necessary dependencies
for the \code{rgee} package, including the specific version of the
\code{earthengine-api} required by the current version of \code{geeLite}. It
is based on the \code{ee_install} function from the \code{rgee} package and
ensures that other essential packages, like \code{geojsonio} and
\code{rnaturalearthdata}, are installed and updated if needed.
}
\examples{
# Example: Creating a Conda environment with 'rgee' dependencies
\dontrun{
  gee_install()
}
}
