% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{gee_message}
\alias{gee_message}
\title{Print Google Earth Engine and Python Environment Information}
\usage{
gee_message(user)
}
\arguments{
\item{user}{[mandatory] (character) Specifies the Google account directory
for which information is displayed.}
}
\description{
Prints information about the Google Earth Engine (GEE) and Python
environment.
}
\keyword{internal}
