% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{gee_message}
\alias{gee_message}
\title{Print Earth Engine and Python Environment Information}
\usage{
gee_message(user = NULL, drive_email = NULL)
}
\arguments{
\item{user}{[optional] (character) Optional Google account identifier to
display.}

\item{drive_email}{[optional] (character) Optional Google Drive email if a
cached token was found. Used only for messaging; can be NULL.}
}
\description{
Prints a summary of the Earth Engine and Python environment used by
  \code{geeLite}, after successful initialization.
}
\keyword{internal}
