% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{get_batches}
\alias{get_batches}
\title{Divide Grid Object into Multiple Batches}
\usage{
get_batches(cases, grid, batch_size)
}
\arguments{
\item{cases}{[mandatory] (integer) Type of data collection request
(\code{1}: All build, \code{2}: All update, \code{3}: Mixed).}

\item{grid}{[mandatory] (sf) Simple features object containing the
geometries of the regions of interest.}

\item{batch_size}{[mandatory] (integer) Maximum size of each batch.}
}
\value{
A list containing two batches:
\describe{
 \item{$b1}{List of primary batches}
 \item{$b2}{List of additional batches (only applicable if 'cases' == 3)}
}
}
\description{
Divides the 'grid' object into multiple batches based on the specified
batch size and the type of data collection request ('cases').
}
\keyword{internal}
