% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{get_config}
\alias{get_config}
\title{Print the Configuration File}
\usage{
get_config(path)
}
\arguments{
\item{path}{[mandatory] (character) The path to the root directory of the
generated database.}
}
\value{
A character string representing the formatted JSON content of the
  configuration file.
}
\description{
Reads and prints the configuration file from the database's root directory
in a human-readable format.
}
\examples{
# Example: Printing the configuration file
\dontrun{
  get_config(path = "path/to/db")
}

}
