% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{get_json}
\alias{get_json}
\title{Print JSON File}
\usage{
get_json(path, file_path)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}

\item{file_path}{[mandatory] (character) Relative path to the JSON file from
the root directory.}
}
\value{
A character string representing the formatted JSON of the specified
file.
}
\description{
Reads and prints a specified JSON file in a human-readable format from the
specified root directory of the generated database.
}
\keyword{internal}
