% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_json.R
\name{get_state}
\alias{get_state}
\title{Print the State File}
\usage{
get_state(path)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}
}
\value{
A character string representing the formatted state JSON.
}
\description{
Reads and prints the state file in a human-readable format.
}
\examples{
# Example: Printing the state file
\dontrun{
  get_state(path = "path/to/root/directory")
}

}
