% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{init_postp}
\alias{init_postp}
\title{Initialize Post-Processing Folder and Files}
\usage{
init_postp(path, verbose = TRUE)
}
\arguments{
\item{path}{[mandatory] \code{character} The path to the root directory
where the \code{postp} folder should be created.}

\item{verbose}{[optional] (logical) Display messages (default: \code{TRUE}).}
}
\description{
Creates a \code{postp} folder at the specified path and adds two empty
files: \code{structure.json} and \code{functions.R}.
}
\details{
The \code{structure.json} file is initialized with a default JSON structure:
\code{"default": null}. This file is intended for mapping variables to
post-processing functions. The \code{functions.R} file is created with a
placeholder comment indicating where to define the R functions for
post-processing. If the \code{postp} folder already exists, an error will
be thrown to prevent overwriting existing files.
}
\examples{
\dontrun{
# Initialize post-processing folder and files in the database root directory
init_postp("/path/to/database")
}
}
