% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_config.R
\name{modify_config}
\alias{modify_config}
\title{Modify Configuration File}
\usage{
modify_config(path, keys, new_values, verbose = TRUE)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}

\item{keys}{[mandatory] (list) Specifying the path to values in the
configuration file that need updating.}

\item{new_values}{[mandatory] (list) New values to replace the original
values specified by 'keys'.}

\item{verbose}{[optional] (logical) Display messages (default: \code{TRUE}).}
}
\description{
Modifies the configuration file in the specified directory of the generated
database (\code{config/config.json}) by updating values associated with the
specified keys.
}
\examples{
# Example: Modifying the configuration file
\dontrun{
  modify_config(
    path = "path/to/db",
    keys = list("limit", c("source", "MODIS/061/MOD13A2", "NDVI")),
    new_values = list(1000, "mean")
  )
}
}
