% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cli.R
\name{process_single_file}
\alias{process_single_file}
\title{Process a Single Source File}
\usage{
process_single_file(src_file_path, path)
}
\arguments{
\item{src_file_path}{[mandatory] (character) The path of the source file to
process.}

\item{path}{[mandatory] (character) The path to the root directory of the
generated database.}
}
\description{
Processes an individual source file by updating the file with the specified
'path' and writing the updated file to the \code{cli/} directory of the
database.
}
\keyword{internal}
