% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{read_db}
\alias{read_db}
\title{Reading Selected SQLite Tables}
\usage{
read_db(path, tables = "all")
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}

\item{tables}{[optional] (character or integer) Names or IDs of the selected
tables. To identify available tables and their names or IDs, use the
\code{get_tables} function (default: \code{"all"}).}
}
\value{
A list object with the first element (grid) as an sf object, and
subsequent elements (databases) as data frame objects.
}
\description{
Reading selected SQLite tables into an \code{list} object.
}
\examples{
# Example: Reading the "grid" table
\dontrun{
  read_db(path = "path", tables = "grid")
}

}
