% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{read_db}
\alias{read_db}
\title{Read Selected SQLite Tables}
\usage{
read_db(path, tables = "all", freq = NULL, temp_stats = NULL)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}

\item{tables}{[optional] (character or integer) Names or IDs of the tables
to be read. Use the \code{fetch_tables} function to identify available table
names and IDs (default: \code{"all"}).}

\item{freq}{[optional] (character) Specifies the frequency to aggregate the
data (options: \code{NULL}, \code{"month"}, \code{"year"}). Default is
\code{NULL}, which means no aggregation.}

\item{temp_stats}{[optional] (character) A character vector of statistical
functions aggregation is based on (options: \code{NULL}, \code{"mean"},
\code{"median"}, \code{"min"}, \code{"max"}, \code{"sd"}). Default is
\code{NULL}, which means no aggregation.}
}
\value{
A list where the first element ('grid') is an simple feature (sf)
object, and subsequent elements are data frame objects.
}
\description{
Reads SQLite tables from the specified directory of the generated database
(\code{data/geelite.db}) into a list object.
}
\examples{
# Example: Reading the 'grid' table
\dontrun{
  db_list <- read_db(path = "path/to/db",
                     tables = "grid")
}

}
