% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{read_tables}
\alias{read_tables}
\title{Read Tables from Database}
\usage{
read_tables(path, tables, freq, prep_fun, aggr_funs)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}

\item{tables}{[mandatory] (character) A vector of table names to be read.}

\item{freq}{[mandatory] (character) Specifies the frequency to aggregate the
data (options: \code{"day"}, \code{"week"}, \code{"month"},
\code{"bimonth"}, \code{"quarter"}, \code{"season"}, \code{"halfyear"},
\code{"year"}).}

\item{prep_fun}{[mandatory] (function) A single function used for
pre-processing the time series data before aggregation. This function
converts the data to a daily frequency and applies any necessary data
transformation or imputation.}

\item{aggr_funs}{[mandatory] (function or list) Specifies the aggregation
function(s) to be applied to the data. This can be a single function or a
list of functions.}
}
\value{
A list of tables read from the database.
}
\description{
Reads the specified tables from the SQLite database.
}
\keyword{internal}
