% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cli.R
\name{set_cli}
\alias{set_cli}
\title{Initialize CLI Files}
\usage{
set_cli(path, verbose = TRUE)
}
\arguments{
\item{path}{[mandatory] (character) The path to the root directory of the
generated database. This must be a writable, non-temporary directory.
Avoid using the home directory (~), the current working directory, or the
package directory.}

\item{verbose}{[optional] (logical) Whether to display messages (default:
\code{TRUE}).}
}
\value{
No return value, called for side effects.
}
\description{
Creates R scripts to enable the main functions to be called through the
Command Line Interface (CLI). These scripts are stored in the \code{cli/}
directory of the generated database.
}
\examples{
\dontrun{
  set_cli(path = tempdir())
}

}
