% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cli.R
\name{set_cli}
\alias{set_cli}
\title{Printing the Configuration File}
\usage{
set_cli(path)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}
}
\value{
Configuration object.
}
\description{
This function prints the configuration file.
}
\examples{
# Example: Printing the configuration file
\dontrun{
  set_cli(path = "path")
}

}
