% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cli.R
\name{set_cli}
\alias{set_cli}
\title{Initialize CLI Files}
\usage{
set_cli(path, verbose = TRUE)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}

\item{verbose}{[optional] (logical) Display messages (default: \code{TRUE}).}
}
\description{
Creates R scripts to enable main functions to be callable through the
Command Line Interface (CLI). These scripts are stored within the specified
directory of the generated database (\code{cli/...}).
}
\examples{
# Example: Setting up CLI files
\dontrun{
  set_cli(path = "path/to/root/directory")
}

}
