% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config.R
\name{set_config}
\alias{set_config}
\title{Setting Up the Configuration File}
\usage{
set_config(
  path,
  regions,
  source,
  start = "2000-01-01",
  resol,
  scale = NULL,
  limit = 10000,
  crs = NULL
)
}
\arguments{
\item{path}{[mandatory] (character) Path to the root directory of the
generated database.}

\item{regions}{[mandatory] (character) FIPS codes of the regions of interest
(two-letter country code, followed by an optional two-digit state code).}

\item{source}{[mandatory] (list) Description of GEE datasets of interest. It
is a nested list with three levels:
\describe{
  \item{names}{(list) Names of the GEE datasets of interest (e.g.,
  \code{"MODIS/061/MOD13A1"}).
    \describe{
      \item{bands}{(list) Bands of interest (e.g., \code{"NDVI"}).
        \describe{
          \item{stats}{(character) Statistics of interest
          (options: \code{"mean"}, \code{"median"}, \code{"min"},
          \code{"max"}, \code{"sd"}).}
        }
      }
    }
  }
}}

\item{start}{[optional] (date) First date of the data collection
(default: \code{"2000-01-01"}).}

\item{resol}{[mandatory] (integer) Resolution of the H3 bin.}

\item{scale}{[optional] (integer) Scale of images before processing
(default: \code{NULL}).}

\item{limit}{[optional] (integer) Limit on concurrent zonal statistics
calculations. \code{Limit - 1} bins will be processed at the same time
(default: \code{10000}).}

\item{crs}{[optional] (integer) CRS to be assigned to the dataset (default
is the shapefile's CRS: \code{NULL}).}
}
\description{
This function creates the configuration file
(\code{path/config/config.json}).
}
\examples{
# Example: Setting up the configuration file
\dontrun{
set_config(path = "path",
           regions = c("SO", "YM"),
           source = list(
            "MODIS/006/MOD13A2" = list(
              "NDVI" = c("mean", "sd")
            )
           ),
           start = "2010-01-01",
           resol = 3)
}

}
