% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config.R
\name{set_config}
\alias{set_config}
\title{Initialize the Configuration File}
\usage{
set_config(
  path,
  regions,
  source,
  start = "2010-01-01",
  resol,
  scale = NULL,
  limit = 10000,
  crs = NULL,
  verbose = TRUE
)
}
\arguments{
\item{path}{[mandatory] (character) The path to the root directory of the
generated database.}

\item{regions}{[mandatory] (character) ISO 3166-2 codes of the regions of
interest (two letters for countries and additional characters for states).}

\item{source}{[mandatory] (list) Description of Google Earth Engine (GEE)
datasets of interest. It is a nested list with three levels:
\describe{
  \item{names}{(list) Datasets of interest (e.g.,
  \code{"MODIS/061/MOD13A1"}).
    \describe{
      \item{bands}{(list) Bands of interest (e.g., \code{"NDVI"}).
        \describe{
          \item{stats}{(character) Statistics of interest
          (options: \code{"mean"}, \code{"median"}, \code{"min"},
          \code{"max"}, \code{"sd"}).}
        }
      }
    }
  }
}}

\item{start}{[optional] (date) First date of the data collection
(default: \code{"2010-01-01"}).}

\item{resol}{[mandatory] (integer) Resolution of the H3 bin.}

\item{scale}{[optional] (integer) Scale of images before processing
(default: \code{NULL}).}

\item{limit}{[optional] (integer) Limit on concurrent zonal statistics
calculations. \code{Limit - 1} bins will be processed at the same time
(default: \code{10000}).}

\item{crs}{[optional] (integer) CRS to be assigned to the dataset (default
is the shapefile's original CRS: \code{NULL}). Note: Reprojection is not
recommended unless essential.}

\item{verbose}{[optional] (logical) Display messages (default: \code{TRUE}).}
}
\description{
Creates the configuration file in the specified directory of the generated
database (\code{config/config.json}).
}
\examples{
# Example: Setting up the configuration file
\dontrun{
  set_config(path = "path/to/root/directory",
             regions = c("SO", "YM"),
             source = list(
              "MODIS/006/MOD13A2" = list(
                "NDVI" = c("mean", "sd")
             )
            ),
            resol = 3)
}
}
