% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R,
%   R/run_geelite._halfgoogleimportR.R
\name{set_depend}
\alias{set_depend}
\title{Set Dependencies}
\usage{
set_depend(conda = "rgee", user = NULL, drive = TRUE, verbose = TRUE)

set_depend(conda = "rgee", user = NULL, drive = TRUE, verbose = TRUE)
}
\arguments{
\item{conda}{[optional] (character) Name of the Conda environment
(default: "rgee").}

\item{user}{[optional] (character) Google account directory under
~/.config/earthengine. Typically NULL unless multiple accounts are used.}

\item{drive}{[optional] (logical) Whether to enable Google Drive access
(default: TRUE).}

\item{verbose}{[optional] (logical) Whether to print a success message at
the end.}
}
\description{
Activates the Conda environment used for Earth Engine, ensures that
authentication exists if needed, and initializes the Earth Engine Python API.

Activates the Conda environment used for Earth Engine, ensures that
authentication exists if needed, and initializes the Earth Engine Python API.
}
\keyword{internal}
