% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{set_depend}
\alias{set_depend}
\title{Set Dependencies}
\usage{
set_depend(conda = "rgee", user = NULL, drive = FALSE, verbose = TRUE)
}
\arguments{
\item{conda}{[optional] (character) Name of the Conda environment
(default: "rgee").}

\item{user}{[optional] (character) Google account directory under
~/.config/earthengine. Typically \code{NULL} unless multiple accounts are
used.}

\item{drive}{[optional] (logical) Whether to enable Google Drive access
(default: \code{FALSE}).}

\item{verbose}{[optional] (logical) Whether to print a success message at
the end.}
}
\description{
Activates the Conda environment used for Earth Engine, ensures
authentication is available if needed, and initializes the Earth Engine
Python API.
}
\keyword{internal}
