% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geelite.R
\name{set_depend}
\alias{set_depend}
\title{Set Dependencies}
\usage{
set_depend(conda = "rgee", user = NULL, drive = FALSE, verbose = TRUE)
}
\arguments{
\item{conda}{[optional] (character) Name of the virtual Conda environment
used by the \code{rgee} package (default: \code{"rgee"}).}

\item{user}{[optional] (character) Specifies the Google account directory
within \code{~/.config/earthengine/}. This directory stores credentials
for a specific Google account (default: \code{NULL}).}

\item{drive}{[optional] (logical) If \code{TRUE}, initializes Google Drive
authentication for tasks involving Drive exports (default: \code{FALSE}).}

\item{verbose}{[optional] (logical) Display messages (default: \code{TRUE}).}
}
\description{
Authenticates the Google Earth Engine (GEE) account and activates the
specified Conda environment.
}
\keyword{internal}
