% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{source_with_notification}
\alias{source_with_notification}
\title{Source an R Script with Notifications About Functions Loaded}
\usage{
source_with_notification(file)
}
\arguments{
\item{file}{[mandatory] (character) A character string specifying the path
to the R script to be sourced.}
}
\value{
An environment containing the functions loaded from the sourced file.
}
\description{
Sources an R script into a dedicated environment and lists the functions
that have been loaded.
}
\keyword{internal}
