% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_db.R
\name{validate_tables_param}
\alias{validate_tables_param}
\title{Validate Tables Parameter}
\usage{
validate_tables_param(tables, tables_all)
}
\arguments{
\item{tables}{[mandatory] (character or integer) A vector specifying tables
to be read.}

\item{tables_all}{[mandatory] (data.frame) All available tables.}
}
\value{
A character vector of valid table names.
}
\description{
Validates the 'tables' parameter and determines which tables to read.
}
\keyword{internal}
