% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_tables_param}
\alias{validate_tables_param}
\title{Validate Tables Parameter}
\usage{
validate_tables_param(tables, tables_all, freq, funs)
}
\arguments{
\item{tables}{[mandatory] (character or integer) A vector specifying tables
to be read.}

\item{tables_all}{[mandatory] (data.frame) All available tables.}

\item{freq}{[mandatory] (character) Specifies the frequency for aggregation
(options: \code{NULL}, \code{"month"}, \code{"year"}).}

\item{funs}{[optional] (character) A character vector of statistical
functions aggregation is based on (options: \code{NULL}, \code{"mean"},
\code{"median"}, \code{"min"}, \code{"max"}, \code{"sd"}).}
}
\value{
A character vector of valid table names.
}
\description{
Validates the 'tables' parameter and determines which tables to read.
}
\keyword{internal}
